(PC-19910): refactor(share): rename shareApp in share
