+++ src/App.tsx
@@ -21,7 +21,7 @@ import { AppNavigationContainer } from 'features/navigation/NavigationContainer'
 import { PushNotificationsWrapper } from 'features/notifications/askNotificationsModal/helpers/PushNotificationsWrapper'
 import { OnboardingWrapper } from 'features/onboarding/context/OnboardingWrapper'
 import { SearchWrapper } from 'features/search/context/SearchWrapper'
-import { ShareAppWrapper } from 'features/shareApp/context/ShareAppWrapper'
+import { ShareAppWrapper } from 'features/share/context/ShareAppWrapper'
 import { initAlgoliaAnalytics } from 'libs/algolia/analytics/initAlgoliaAnalytics'
 import { SearchAnalyticsWrapper } from 'libs/algolia/analytics/SearchAnalyticsWrapper'
 import { campaignTracker } from 'libs/campaign'
