+++ src/features/auth/pages/signup/AccountCreated/AccountCreated.native.test.tsx
@@ -5,8 +5,8 @@ import { navigate } from '__mocks__/@react-navigation/native'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { navigateFromRef } from 'features/navigation/navigationRef'
-import { ShareAppWrapper } from 'features/shareApp/context/ShareAppWrapper'
-import { ShareAppModalType } from 'features/shareApp/helpers/shareAppModalInformations'
+import { ShareAppWrapper } from 'features/share/context/ShareAppWrapper'
+import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
 import { BatchUser } from 'libs/react-native-batch'
 import { render, fireEvent } from 'tests/utils'
