+++ src/features/identityCheck/pages/confirmation/BeneficiaryAccountCreated.tsx
@@ -9,8 +9,8 @@ import {
 import { navigateToHome, navigateToHomeConfig } from 'features/navigation/helpers'
 import { isUserUnderageBeneficiary } from 'features/profile/helpers/isUserUnderageBeneficiary'
 import { useMaxPrice } from 'features/search/helpers/useMaxPrice/useMaxPrice'
-import { useShareAppContext } from 'features/shareApp/context/ShareAppWrapper'
-import { ShareAppModalType } from 'features/shareApp/helpers/shareAppModalInformations'
+import { useShareAppContext } from 'features/share/context/ShareAppWrapper'
+import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
 import { formatPriceInEuroToDisplayPrice } from 'libs/parsers'
 import { BatchEvent, BatchUser } from 'libs/react-native-batch'
 import TutorialPassLogo from 'ui/animations/tutorial_pass_logo.json'
