+++ src/features/share/helpers/shareApp.ts
