+++ src/features/share/helpers/shareAppModalInformations.ts
