+++ src/libs/firebase/analytics/analytics.ts
@@ -6,7 +6,7 @@ import { FavoriteSortBy } from 'features/favorites/types'
 import { IdentityCheckStep } from 'features/identityCheck/types'
 import { Referrals } from 'features/navigation/RootNavigator/types'
 import { SearchState } from 'features/search/types'
-import { ShareAppModalType } from 'features/shareApp/helpers/shareAppModalInformations'
+import { ShareAppModalType } from 'features/share/helpers/shareAppModalInformations'
 import { ContentTypes } from 'libs/contentful'
 import { AnalyticsEvent } from 'libs/firebase/analytics/events'
 import { analyticsProvider } from 'libs/firebase/analytics/provider'
